Revert "Feature/single room page"
